feat: add AgentMessage struct for agent communication
