+++ crates/forge_domain/src/orch.rs
@@ -5,7 +5,6 @@ use anyhow::Context as AnyhowContext;
 use async_recursion::async_recursion;
 use futures::future::join_all;
 use futures::{Stream, StreamExt};
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use tokio::sync::RwLock;
 use tokio_retry::strategy::{jitter, ExponentialBackoff};
@@ -19,17 +18,7 @@ use crate::*;
 
 type ArcSender = Arc<tokio::sync::mpsc::Sender<anyhow::Result<AgentMessage<ChatResponse>>>>;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentMessage<T> {
-    pub agent: AgentId,
-    pub message: T,
-}
-
-impl<T> AgentMessage<T> {
-    pub fn new(agent: AgentId, message: T) -> Self {
-        Self { agent, message }
-    }
-}
+use crate::agent::AgentMessage;
 
 #[derive(Clone)]
 pub struct Orchestrator<App> {
